+++ api/src/main/java/org/openmrs/api/db/hibernate/HibernateContextDAO.java
@@ -17,25 +17,19 @@ import java.io.InputStream;
 import java.util.Map;
 import java.util.Properties;
 
-import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.hibernate.FlushMode;
-import org.hibernate.Hibernate;
-import org.hibernate.HibernateException;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.stat.QueryStatistics;
 import org.hibernate.stat.Statistics;
+import org.hibernate.type.StringType;
 import org.hibernate.util.ConfigHelper;
-import org.openmrs.GlobalProperty;
 import org.openmrs.User;
 import org.openmrs.api.context.Context;
-import org.openmrs.api.context.ContextAuthenticationException;
 import org.openmrs.api.db.ContextDAO;
-import org.openmrs.util.OpenmrsConstants;
 import org.openmrs.util.OpenmrsUtil;
-import org.openmrs.util.Security;
 import org.springframework.orm.hibernate3.SessionFactoryUtils;
 import org.springframework.orm.hibernate3.SessionHolder;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
@@ -67,136 +61,6 @@ public class HibernateContextDAO implements ContextDAO {
 	}
 	
 	/**
-	 * @see org.openmrs.api.db.ContextDAO#authenticate(java.lang.String, java.lang.String)
-	 */
-	public User authenticate(String login, String password) throws ContextAuthenticationException {
-		
-		String errorMsg = "Invalid username and/or password: " + login;
-		
-		Session session = sessionFactory.getCurrentSession();
-		
-		User candidateUser = null;
-		
-		if (login != null) {
-			//if username is blank or white space character(s)
-			if (StringUtils.isEmpty(login) || StringUtils.isWhitespace(login))
-				throw new ContextAuthenticationException(errorMsg);
-			
-			// loginWithoutDash is used to compare to the system id
-			String loginWithDash = login;
-			if (login.matches("\\d{2,}"))
-				loginWithDash = login.substring(0, login.length() - 1) + "-" + login.charAt(login.length() - 1);
-			
-			try {
-				candidateUser = (User) session.createQuery(
-				    "from User u where (u.username = ? or u.systemId = ? or u.systemId = ?) and u.retired = '0'").setString(
-				    0, login).setString(1, login).setString(2, loginWithDash).uniqueResult();
-			}
-			catch (HibernateException he) {
-				log.error("Got hibernate exception while logging in: '" + login + "'", he);
-			}
-			catch (Exception e) {
-				log.error("Got regular exception while logging in: '" + login + "'", e);
-			}
-		}
-		
-		// only continue if this is a valid username and a nonempty password
-		if (candidateUser != null && password != null) {
-			if (log.isDebugEnabled())
-				log.debug("Candidate user id: " + candidateUser.getUserId());
-			
-			String lockoutTimeString = candidateUser.getUserProperty(OpenmrsConstants.USER_PROPERTY_LOCKOUT_TIMESTAMP, null);
-			Long lockoutTime = null;
-			if (lockoutTimeString != null && !lockoutTimeString.equals("0")) {
-				try {
-					// putting this in a try/catch in case the admin decided to put junk into the property
-					lockoutTime = Long.valueOf(lockoutTimeString);
-				}
-				catch (NumberFormatException e) {
-					log.debug("bad value stored in " + OpenmrsConstants.USER_PROPERTY_LOCKOUT_TIMESTAMP + " user property: "
-					        + lockoutTimeString);
-				}
-			}
-			
-			// if they've been locked out, don't continue with the authentication
-			if (lockoutTime != null) {
-				// unlock them after 5 mins, otherwise reset the timestamp
-				// to now and make them wait another 5 mins
-				if (System.currentTimeMillis() - lockoutTime > 300000) {
-					candidateUser.setUserProperty(OpenmrsConstants.USER_PROPERTY_LOGIN_ATTEMPTS, "0");
-					candidateUser.removeUserProperty(OpenmrsConstants.USER_PROPERTY_LOCKOUT_TIMESTAMP);
-					saveUserProperties(candidateUser);
-				} else {
-					candidateUser.setUserProperty(OpenmrsConstants.USER_PROPERTY_LOCKOUT_TIMESTAMP, String.valueOf(System
-					        .currentTimeMillis()));
-					throw new ContextAuthenticationException(
-					        "Invalid number of connection attempts. Please try again later.");
-				}
-			}
-			
-			String passwordOnRecord = (String) session.createSQLQuery("select password from users where user_id = ?")
-			        .addScalar("password", Hibernate.STRING).setInteger(0, candidateUser.getUserId()).uniqueResult();
-			
-			String saltOnRecord = (String) session.createSQLQuery("select salt from users where user_id = ?").addScalar(
-			    "salt", Hibernate.STRING).setInteger(0, candidateUser.getUserId()).uniqueResult();
-			
-			// if the username and password match, hydrate the user and return it
-			if (passwordOnRecord != null && Security.hashMatches(passwordOnRecord, password + saltOnRecord)) {
-				// hydrate the user object
-				candidateUser.getAllRoles().size();
-				candidateUser.getUserProperties().size();
-				candidateUser.getPrivileges().size();
-				
-				// only clean up if the were some login failures, otherwise all should be clean
-				Integer attempts = getUsersLoginAttempts(candidateUser);
-				if (attempts > 0) {
-					candidateUser.setUserProperty(OpenmrsConstants.USER_PROPERTY_LOGIN_ATTEMPTS, "0");
-					candidateUser.removeUserProperty(OpenmrsConstants.USER_PROPERTY_LOCKOUT_TIMESTAMP);
-					saveUserProperties(candidateUser);
-				}
-				
-				// skip out of the method early (instead of throwing the exception)
-				// to indicate that this is the valid user
-				return candidateUser;
-			} else {
-				// the user failed the username/password, increment their
-				// attempts here and set the "lockout" timestamp if necessary
-				Integer attempts = getUsersLoginAttempts(candidateUser);
-				
-				attempts++;
-				
-				Integer allowedFailedLoginCount = 7;
-				
-				try {
-					allowedFailedLoginCount = Integer.valueOf(Context.getAdministrationService().getGlobalProperty(
-					    OpenmrsConstants.GP_ALLOWED_FAILED_LOGINS_BEFORE_LOCKOUT).trim());
-				}
-				catch (Exception ex) {
-					log.error("Unable to convert the global property "
-					        + OpenmrsConstants.GP_ALLOWED_FAILED_LOGINS_BEFORE_LOCKOUT
-					        + "to a valid integer. Using default value of 7");
-				}
-				
-				if (attempts > allowedFailedLoginCount) {
-					// set the user as locked out at this exact time
-					candidateUser.setUserProperty(OpenmrsConstants.USER_PROPERTY_LOCKOUT_TIMESTAMP, String.valueOf(System
-					        .currentTimeMillis()));
-				} else {
-					candidateUser.setUserProperty(OpenmrsConstants.USER_PROPERTY_LOGIN_ATTEMPTS, String.valueOf(attempts));
-				}
-				
-				saveUserProperties(candidateUser);
-			}
-		}
-		
-		// throw this exception only once in the same place with the same
-		// message regardless of username/pw combo entered
-		log.info("Failed login attempt (login=" + login + ") - " + errorMsg);
-		throw new ContextAuthenticationException(errorMsg);
-		
-	}
-	
-	/**
 	 * @see org.openmrs.api.db.ContextDAO#getUserByUuid(java.lang.String)
 	 */
 	public User getUserByUuid(String uuid) {
@@ -212,33 +76,6 @@ public class HibernateContextDAO implements ContextDAO {
 		sessionFactory.getCurrentSession().setFlushMode(flushMode);
 		
 		return u;
-	}
-	
-	/**
-	 * Call the UserService to save the given user while proxying the privileges needed to do so.
-	 * 
-	 * @param user the User to save
-	 */
-	private void saveUserProperties(User user) {
-		sessionFactory.getCurrentSession().update(user);
-	}
-	
-	/**
-	 * Get the integer stored for the given user that is their number of login attempts
-	 * 
-	 * @param user the user to check
-	 * @return the # of login attempts for this user defaulting to zero if none defined
-	 */
-	private Integer getUsersLoginAttempts(User user) {
-		String attemptsString = user.getUserProperty(OpenmrsConstants.USER_PROPERTY_LOGIN_ATTEMPTS, "0");
-		Integer attempts = 0;
-		try {
-			attempts = Integer.valueOf(attemptsString);
-		}
-		catch (NumberFormatException e) {
-			// skip over errors and leave the attempts at zero
-		}
-		return attempts;
 	}
 	
 	/**
@@ -398,6 +235,22 @@ public class HibernateContextDAO implements ContextDAO {
 			}
 		}
 		
+	}
+	
+	/**
+	 * @see org.openmrs.api.db.ContextDAO#getHashPassword(org.openmrs.User)
+	 */
+	public String getHashedPassword(User user) {
+		return (String) sessionFactory.getCurrentSession().createSQLQuery("select password from users where user_id = ?")
+		        .addScalar("password", StringType.INSTANCE).setInteger(0, user.getUserId()).uniqueResult();
+	}
+	
+	/**
+	 * @see org.openmrs.api.db.ContextDAO#getSalt(org.openmrs.User)
+	 */
+	public String getSalt(User user) {
+		return (String) sessionFactory.getCurrentSession().createSQLQuery("select salt from users where user_id = ?")
+		        .addScalar("salt", StringType.INSTANCE).setInteger(0, user.getUserId()).uniqueResult();
 	}
 	
 }
